plus_sygma_pub_export/requirements.txt
python-dotenv==0.10.1
--extra-index-url https://testpypi.python.org/pypi
openid-http-client==0.0.22